+++ com.unity.render-pipelines.lightweight/Runtime/Data/LightweightPipelineEditorResources.cs
@@ -1,0 +1,12 @@
+using System;
+using UnityEngine;
+
+public class LightweightPipelineEditorResources : ScriptableObject
+{
+    public Material DefaultMaterial;
+    public Material DefaultParticleMaterial;
+    public Material DefaultTerrainMaterial;
+    public Shader AutodeskInteractiveShader;
+    public Shader AutodeskInteractiveTransparentShader;
+    public Shader AutodeskInteractiveMaskedShader;
+}