+++ com.unity.shadergraph/Editor/Data/Util/GraphUtil.cs
@@ -1021,11 +1021,11 @@ namespace UnityEditor.ShaderGraph
 
                 finalShader.AppendLine(@"HLSLINCLUDE");
                 finalShader.AppendLine("#define USE_LEGACY_UNITY_MATRIX_VARIABLES");
-                finalShader.AppendLine(@"#include ""CoreRP/ShaderLibrary/Common.hlsl""");
-                finalShader.AppendLine(@"#include ""CoreRP/ShaderLibrary/Packing.hlsl""");
-                finalShader.AppendLine(@"#include ""CoreRP/ShaderLibrary/Color.hlsl""");
-                finalShader.AppendLine(@"#include ""CoreRP/ShaderLibrary/UnityInstancing.hlsl""");
-                finalShader.AppendLine(@"#include ""CoreRP/ShaderLibrary/EntityLighting.hlsl""");
+                finalShader.AppendLine(@"#include ""Packages/com.unity.render-pipelines.core/ShaderLibrary/Common.hlsl""");
+                finalShader.AppendLine(@"#include ""Packages/com.unity.render-pipelines.core/ShaderLibrary/Packing.hlsl""");
+                finalShader.AppendLine(@"#include ""Packages/com.unity.render-pipelines.core/ShaderLibrary/Color.hlsl""");
+                finalShader.AppendLine(@"#include ""Packages/com.unity.render-pipelines.core/ShaderLibrary/UnityInstancing.hlsl""");
+                finalShader.AppendLine(@"#include ""Packages/com.unity.render-pipelines.core/ShaderLibrary/EntityLighting.hlsl""");
                 finalShader.AppendLine(@"#include ""ShaderGraphLibrary/ShaderVariables.hlsl""");
                 finalShader.AppendLine(@"#include ""ShaderGraphLibrary/ShaderVariablesFunctions.hlsl""");
                 finalShader.AppendLine(@"#include ""ShaderGraphLibrary/Functions.hlsl""");