+++ com.unity.shadergraph/CHANGELOG.md
@@ -59,3 +59,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Normal Create` node has been renamed to `Normal From Texture`.
 - The preview of nodes now updates correctly.
 - Your system locale can no longer cause incorrect commands due to full stops being converted to commas.
+- Texture type input slots now set correct default values for 'Normal' texture type.